initiate commit

+++ app.py
@@ -80,7 +80,9 @@ def name():
 
 @app.route('/delete/id=<int:sid>', methods =["DELETE"])
 def delete(sid):
-	ds	= sensor.query.filter_by(id=sid).first()
+	if request.method != "DELETE":
+		return "Method not allowed, use DELETE method"
+	ds = Sensor.query.filter_by(id=sid).first()
 	if not ds:
 		return "Sensor with id {} not found".format(sid)
 	db.session.delete(ds)
@@ -90,16 +92,16 @@ def delete(sid):
 
 @app.route('/read', methods = ['POST', 'GET'])
 def read():
-	val	= request.form.get('s1')
-	sid	= request.form.get('id')
-	tsp	= datetime.now()
+#	val	= request.form.get('s1')
+#	sid	= request.form.get('id')
+#	tsp	= datetime.now()
 	idc = sensor.query.filter_by(id=sid).first()
 	html_string = "<html><h2>Sensor1 : {}%</h2><h2>ID : {}</h2></html>".format(val,idc)
-	if not idc:
-		return "Device is not registered. Can't send value."
-	dat	= data(tsp=tsp, val=val, sid=sid)
-	db.session.add(dat)
-	db.session.commit()
+#	if not idc:
+#		return "Device is not registered. Can't send value."
+#	dat	= data(tsp=tsp, val=val, sid=sid)
+#	db.session.add(dat)
+#	db.session.commit()
 	return html_string
 
 if __name__ == "__main__":
